Desktop reorg
Major reorganization of directory layouts for ease of cloning.
This should allow the powershell scripts to work out of the box.
A hugo directory is included to showcase where hugo.exe should be placed. Environment variable setting (in Windows e.g.) should not be required.
Only changes are to powershell scripts to account for the website rename.