Link fix-ups
Adjusted links to make all links internal.
For pages that this would cause issues, e.g. Search and Settings, external links were simply removed.
(Note that we may look into "saving" search queries as part of the URL hash, just so users can go back to search results after opening a result. Still working on the back-end!)
Also fixed a little oopsie on the settings page (the Preview box had a h1 header, which gained a <a> link it should not have - changed that to a h2 header instead!)